calc_physical_separation: Aida Behmard bugnote!
The paper should have read: TOI 837 to "Star A" separation of 6.6 +/-
2.1 pc.

Not +/- 0.1 pc.

The DR2 + EDR3 data therefore suggest that "Star A" is probably
physically bound.

~300 AU apparent separation, with highly uncertain parallaxes, but Star
A moving from behind TOI837 (DR2) to in front (EDR3), are all consistent
with it being more than just a nearby member of the cluster, and more
likely a bound companion.